Commit 3
Fixed bug with model loading.
Added detect.cpp
Implemented modelcomponents function.

+++ src/Box.h
@@ -14,7 +14,7 @@ public:
 	float s;
 	float c;
 	Mat xy;
-	double level;
+	float level;
 	Box(void);
 	~Box(void);
 };
